+++ hawkey/src/python/hawkeymodule.c
@@ -147,7 +147,6 @@ init_hawkey(void)
     PyModule_AddIntConstant(m, "PKG_LATEST", HY_PKG_LATEST);
     PyModule_AddIntConstant(m, "PKG_DOWNGRADES", HY_PKG_DOWNGRADES);
     PyModule_AddIntConstant(m, "PKG_UPGRADES", HY_PKG_UPGRADES);
-    PyModule_AddIntConstant(m, "PKG_OBSOLETING", HY_PKG_OBSOLETING);
 
     PyModule_AddIntConstant(m, "CHKSUM_MD5", HY_CHKSUM_MD5);
     PyModule_AddIntConstant(m, "CHKSUM_SHA1", HY_CHKSUM_SHA1);