+++ hawkey/src/python/query-py.h
@@ -7,7 +7,6 @@
 enum py_key_name_e {
     HY_PKG_DOWNGRADES = 100,
     HY_PKG_LATEST,
-    HY_PKG_OBSOLETING,
     HY_PKG_UPGRADES
 };